Stop using the -P (service port) ncsvc option.  Fixes GitHub Issue #3.
Although I cannot find any documentation on this, log results and some
user experiences appear to show that the service port is intended to be
a LOCAL port, not the port on the server, which is what I was using it
for.

I've removed that portion of the code, so we remember the port provided
to use in the server URL (or use 443 if not given) and we test it to
be sure the server is valid.  But, we don't pass this value to ncsvc
with the -P option.

I created a new config value "svcport" that will be sent with -P but
currently there's no way from the UI to set that value.  I'm not sure
what use it is.  Anecdotal evidence seems to suggest that even if
the default (which is 443) is already used, you can still run ncsvc
without specifying -P.